+++ src/args.rs
@@ -1,139 +1,141 @@
-use lazy_static::lazy_static;
-use regex::Regex;
-use crate::help;
+use clap::{Arg, App, ArgGroup, ValueHint, ArgMatches};
 
-lazy_static! {
-    static ref ARG_REGEX: Regex = Regex::new(r#"[^\s"']+|"([^"]*)"|'([^']*)'"#).unwrap();
-}
+pub fn get_app() -> App<'static> {
 
-#[derive(Debug)]
-pub struct MushtArgs {
-    pub mosh_port: String,
-    pub ssh_port: String,
-    pub ssh_args: String,
-    pub user: String,
-    pub host: String,
-    pub args: Vec<String>,
-    pub handled: bool
-}
+    let app: App<'static> = App::new(env!("CARGO_PKG_NAME"))
+        .version(env!("CARGO_PKG_VERSION"))
+        .author(env!("CARGO_PKG_AUTHORS"))
+        .about(concat!(
+            env!("CARGO_PKG_DESCRIPTION"),
+            "\nhomepage: ", env!("CARGO_PKG_HOMEPAGE")))
+        .override_usage("Usage: mosh [options] [--] [user@]host[:ssh-port]")
 
-impl Default for MushtArgs {
-    fn default() -> Self {
-        MushtArgs {
-            mosh_port: String::new(),
-            ssh_port: String::new(),
-            ssh_args: "ssh".to_string(),
-            user: String::new(),
-            host: String::new(),
-            args: Vec::new(),
-            handled: false
-        }
-    }
-}
+        // arg for user/host/port
+        .arg(Arg::new("HOST")
+            .about("the host to connect to")
+            .required(true)
+            .hidden(true)
+            .value_hint(ValueHint::Hostname)
+            .index(1))
+        
+        //args for execs
+        .arg(Arg::new("MOSH")
+            .long("mosh")
+            .value_name("path")
+            .about("mosh executable on local machine")
+            .default_value("mosh")
+            .help_heading(Some("MOSH EXECUTABLES"))
+            .display_order(0))
+        .arg(Arg::new("CLIENT")
+            .long("client")
+            .value_name("path")
+            .about("mosh client on local machine")
+            .default_value("mosh-client")
+            .help_heading(Some("MOSH EXECUTABLES"))
+            .display_order(1))
+        .arg(Arg::new("SERVER")
+            .long("server")
+            .value_name("remote-path")
+            .about("mosh server on remote machine")
+            .default_value("mosh-server")
+            .help_heading(Some("MOSH EXECUTABLES"))
+            .display_order(2))
 
-mod host;
-impl MushtArgs {
-    pub fn get_ssh_args(self: &Self) -> String {
-        if self.ssh_port == "" {
-            self.ssh_args.clone()
-        } else {
-            format!("{} -p {}", self.ssh_args, self.ssh_port)
-        }
-    }
-    pub fn get_user_host(self: &Self) -> String {
-        if self.user == "" {
-            self.host.clone()
-        } else {
-            format!("{}@{}", self.user, self.host)
-        }
-    }
-}
+        //predict arg
+        .arg(Arg::new("PREDICT")
+            .long("predict")
+            .value_name("method")
+            .about("change prediction type \
+            \n      adaptive - local echo for slower links [default] \
+            \n      always - use local echo even on fast links \
+            \n      never - never use local echo \
+            \n      experimental - aggressively echo even when incorrect")
+            .default_value("adaptive")
+            .possible_values(&[ "adaptive", "always", "never", "experimental" ])
+            .hide_possible_values(true)
+            .hide_default_value(true)
+            .help_heading(Some("PREDICTIONS"))
+            .display_order(3))
+        .arg(Arg::new("OVERWRITE")
+            .short('o')
+            .long("predict-overwrite")
+            .about("prediction overwrites instead of inserting")
+            .help_heading(Some("PREDICTIONS"))
+            .display_order(4))
 
-//TODO catch bad arguments
-pub fn parse(mut args: Vec<String>) -> MushtArgs {
-    // give a little wiggle room for expanding vec without relocation
-    let mut musht_args = MushtArgs{args: Vec::with_capacity(args.len()+2), ..Default::default()};
+        //networking args
+        .arg(Arg::new("FAMILY")
+            .long("family")
+            .value_name("method")
+            .about("change ip options \
+            \n      inet - use IPv4 only \
+            \n      inet6 - use IPv6 only \
+            \n      auto - autodetect network type for single-family hosts only \
+            \n      all - try all network types \
+            \n      prefer-inet - use all network types, but try IPv4 first [default] \
+            \n      prefer-inet6 - use all network types, but try IPv6 first ")
+            .default_value("prefer-inet")
+            .possible_values(&[ "inet", "inet6", "auto", "all", "prefer-inet", "prefer-inet6" ])
+            .hide_possible_values(true)
+            .hide_default_value(true)
+            .help_heading(Some("NETWORK OPTIONS"))
+            .display_order(5))
+        .arg(Arg::new("PORT")
+            .short('p')
+            .long("port")
+            .value_name("port[:port2]")
+            .about("server-side UDP port or range (No effect on server-side SSH port)")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .display_order(6))
+        .arg(Arg::new("SSH PORT")
+            .short('s')
+            .long("ssh-port")
+            .value_name("port")
+            .about("server-side TCP ssh port (overrides port specified with --ssh)")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .display_order(7))
+        .arg(Arg::new("SSH")
+            .long("ssh")
+            .value_name("command")
+            .about("ssh command to run when setting up session \
+            \n      [example: \"ssh -p 2222\"]")
+            .default_value("ssh")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .display_order(8))
+        .arg(Arg::new("BIND SERVER")
+            .long("bind-server")
+            .value_name("option")
+            .about("ask the server to reply from an IP address")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .default_value("ssh")
+            .possible_values(&[ "ssh", "any", "IP" ])
+            .display_order(9))
+        .arg(Arg::new("LOCAL")
+            .long("local")
+            .about("run mosh-server locally without using ssh")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .display_order(10))
+        .arg(Arg::new("EXPERIMENTAL REMOTE IP")
+            .long("experimental-remote-ip")
+            .value_name("option")
+            .about("select the method for discovering the remote IP address to use for mosh \n")
+            .help_heading(Some("NETWORK OPTIONS"))
+            .default_value("proxy")
+            .possible_values(&[ "local", "remote", "proxy" ])
+            .display_order(11))
 
-    // get vec of args parsed with spaces, not equal signs
-    for i in 1..args.len() {
+        //client options args
+        .arg(Arg::new("NO INIT")
+            .long("no-init")
+            .about("do not send terminal initialization string")
+            .help_heading(Some("CLIENT OPTIONS"))
+            .display_order(12))
+        .arg(Arg::new("NO SSH PTY")
+            .long("no-ssh-pty")
+            .about("do not allocate a pseudo tty on ssh connection")
+            .help_heading(Some("CLIENT OPTIONS"))
+            .display_order(13));
+    
+    app
 
-        //skip if null arg
-        if args[i] == "" {continue;}
-
-        let full_arg = (*args[i]).to_string().to_lowercase();
-        let is_arg = full_arg.starts_with("-");
-        //make an option be split at space or equals
-        let mut option: Vec<&str> = if is_arg {full_arg.splitn(2, "=").collect()}
-            else {vec!(&full_arg)};
-        let mut next_arg = String::new();
-        if option.len() == 1 && is_arg && i+1 < args.len() {
-            next_arg = args[i+1].clone();
-            option.push(&next_arg);
-            //null arg
-            args[i+1] = String::new();
-        }
-        let mut add_option = true;
-        for arg in &option {
-
-            match *arg {
-                "-p" | "--p" | "-port" | "--port" => {
-                    if option.len() == 2 {
-                        musht_args.mosh_port = option[1].to_string();
-                        add_option = false;
-                    } else if i+1 < args.len() {
-                        musht_args.mosh_port = (*args[i+1]).to_string();
-                    }
-                },
-                "-ssh" | "--ssh" => {
-                    let ssh_args = parse_ssh_args(option[1]);
-                    musht_args.ssh_args = ssh_args.0;
-                    if ssh_args.1 != "" {musht_args.ssh_port = ssh_args.1;}
-                    add_option = false;
-                },
-                "-h" | "--h" | "-help" | "--help" => {
-                    help::print_help();
-                    musht_args.handled = true;
-                },
-                "-v" | "--v" | "-version" | "--version" => {
-                    help::print_version();
-                    musht_args.handled = true;
-                }
-                _ => {
-                    if add_option && i>=1 && !args[i].starts_with("-") && 
-                    (!args[i-1].starts_with("-") || args[i-1].contains("=")) {
-                        let mut userhost: Vec<&str> = arg.splitn(2, "@").collect();
-                        if userhost.len() == 1 {userhost.insert(0, "");}
-                        let hostport: Vec<&str> = userhost[1].splitn(2, ":").collect();
-                        musht_args.user = userhost[0].to_string();
-                        musht_args.host = hostport[0].to_string();
-                        if hostport.len() == 2 {musht_args.ssh_port = hostport[1].to_string()}
-                        add_option = false;
-                    }
-                }
-            }
-
-            if add_option {musht_args.args.push((*arg).to_string())};
-        }
-    }
-
-    musht_args
-
-}
-
-fn parse_ssh_args(ssh_args: &str) -> (String, String){
-    let mut args: Vec<&str> = ARG_REGEX.find_iter(ssh_args).map(|x| x.as_str()).collect();
-
-    let mut port = "";
-    for i in 0..args.len() {
-        if args[i].contains("-p") {
-            port = &args[i][2..];
-            if port == "" && i+1 < args.len() {
-                port = args[i+1];
-                args[i+1] = "";
-            }
-            //now efectively remove from args
-            args[i] = "";
-        }
-    }
-    (args.into_iter().filter(|x| *x != "").collect::<Vec<&str>>().join(" "), port.to_string())
 }
